requirements.txt
pillow
sv-ttk
pywinstyles
tkinterdnd2
pyinstaller